+++ example/lib/base/config.dart
@@ -8,6 +8,9 @@ class Config {
   static bool isWeb = true;     //是否是Web运行   【打包APP时改为false】
   static bool isTest = false;   //是否是测试环境   【不用修改 Native调用程序自己会判断】
 
+  static String appBarHeight = "";              //导航栏+状态栏总高度
+
+
   static String token = "";              //当前Token
   static String userId = "";             //登录用户ID
   static String serverAddress = "http://ymbok.com";