+++ example/lib/pages/demo/error.dart
@@ -69,16 +69,16 @@ class _ErrorPageState extends State<ErrorPage> with SingleNativeStateMixin{
               ),
 
               isLoading?Positioned(
-                  top: 80,
+                  top: appBarHeight,
                   left:0,
                   width:  MediaQuery.of(context).size.width,
-                  height:MediaQuery.of(context).size.height - 80,
+                  height:MediaQuery.of(context).size.height - appBarHeight,
                   child:YmLoading())
                   :Positioned(
-                top: 80,
+                top: MediaQuery.of(context).padding.top + 56,
                 left:0,
                 width:  MediaQuery.of(context).size.width,
-                height:MediaQuery.of(context).size.height - 80,
+                height:MediaQuery.of(context).size.height - appBarHeight,
                 child:YmError("网络状态不佳","assets/images/ic_no_data.png")
               ),