+++ example/lib/pages/demo/loading.dart
@@ -1,3 +1,4 @@
+import 'package:example/base/single_native_state_mixin.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:example/base/config.dart';
@@ -17,14 +18,7 @@ class LoadingPage extends StatefulWidget {
   _LoadingPageState createState() => _LoadingPageState();
 }
 
-class _LoadingPageState extends State<LoadingPage> {
-
-  //和Native进行通信
-  static const flutterChannel = const MethodChannel("sample.flutter.io/flutter");
-  static const nativeChannel = const MethodChannel("sample.flutter.io/native");
-
-  bool _isRouteFromFlutter = false;
-  bool _isLoading = true;
+class _LoadingPageState extends State<LoadingPage> with SingleNativeStateMixin{
 
   @override
   void initState() {
@@ -35,40 +29,19 @@ class _LoadingPageState extends State<LoadingPage> {
       dynamic  arguments = ModalRoute.of(context)!.settings.arguments;
       print("路由传递过来的参数：" + arguments.toString());
       if (arguments != null) {
-        _isRouteFromFlutter = arguments["flutter"];
-        if(Config.isWeb || _isRouteFromFlutter){
+        isRouteFromFlutter = arguments["flutter"];
+        if(Config.isWeb || isRouteFromFlutter){
           //加载数据
+          Future.delayed(Duration(seconds: 1), () {
+            setState(() {
+              isLoading = false;
+            });
+          });
+
         }
       }
     });
-
-
-    Future.delayed(Duration(seconds: 1), () {
-      setState(() {
-        _isLoading = false;
-      });
-
-    });
-
   }
-
-  ///Flutter调用原生
-  Future<void> _goBack() async{
-    if(_isRouteFromFlutter){
-      //如果是Flutter导航进来的页面直接退出
-      Navigator.pop(context);
-    }else{
-      //如果该页面是Native直接打开的，发送完成消息给Native处理
-      try {
-        Map<String, dynamic> arguments = {'message': 'onFinish'};
-        String result = await flutterChannel.invokeMethod('callNative', arguments);
-        print(result);
-      } on PlatformException catch (e) {
-        print("Failed: '${e.message}'.");
-      }
-    }
-  }
-
 
   @override
   Widget build(BuildContext context) {
@@ -86,21 +59,21 @@ class _LoadingPageState extends State<LoadingPage> {
                   child: YmAppBar(widget.title,[
                     const Color(0xFF606FFF),
                     const Color(0xFF3446F2),
-                  ],_goBack,titleTextColor: Colors.white,),
+                  ],goBack,titleTextColor: Colors.white,),
                 ),
               ),
 
-              _isLoading?Positioned(
-                  top: 80,
+              isLoading?Positioned(
+                  top: appBarHeight,
                   left:0,
                   width:  MediaQuery.of(context).size.width,
-                  height:MediaQuery.of(context).size.height - 80,
+                  height:MediaQuery.of(context).size.height - appBarHeight,
                   child:YmLoading())
                   :Positioned(
                       top: 80,
                       left:0,
                       width:  MediaQuery.of(context).size.width,
-                      height:MediaQuery.of(context).size.height - 80,
+                      height:MediaQuery.of(context).size.height - appBarHeight,
                       child:Center(
                         child:Text("页面加载完成") ,
                       ),