+++ example/lib/pages/demo/toast.dart
@@ -59,7 +59,7 @@ class _ToastPageState extends State<ToastPage> with SingleNativeStateMixin{
             ),
 
             Positioned(
-              top: 80,
+              top: appBarHeight,
               left:0,
               width:  MediaQuery.of(context).size.width,
               child:Padding(