+++ example/lib/pages/demo/video.dart
@@ -89,7 +89,7 @@ class _VideoPageState extends State<VideoPage> with SingleNativeStateMixin{
             ),
 
             Positioned(
-              top: 80,
+              top: appBarHeight,
               left:0,
               width:  MediaQuery.of(context).size.width,
               child:Padding(