+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.5] - 2021/06/11.
+* add ListView LoadMore Widget and Sample
+* add Dialog Box Widget and update Cascader Sample with Dialog Box Widget
+
 ## [0.0.4] - 2021/06/04.
 * add TabPageView Widget
 * add TabPageView Sample